[BACKLOG-17443] More reviews
