+++ pentaho-webpackage/assemblies/pentaho-webpackage-archive-deployer/pom.xml
@@ -9,22 +9,15 @@
     <version>8.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>pentaho-webpackage</artifactId>
+  <artifactId>pentaho-webpackage-archive-deployer</artifactId>
 
   <packaging>feature</packaging>
 
   <dependencies>
     <dependency>
       <groupId>org.pentaho.webpackage</groupId>
-      <artifactId>http-extender-impl</artifactId>
+      <artifactId>archive-deployer-impl</artifactId>
       <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.pentaho.webpackage</groupId>
-      <artifactId>pentaho-webpackage-archive-deployer</artifactId>
-      <version>${project.version}</version>
-      <type>xml</type>
-      <classifier>features</classifier>
     </dependency>
   </dependencies>
 </project>
