--- pentaho-webpackage/assemblies/pentaho-webpackage-archive-file-deployer/pom.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.pentaho.webpackage</groupId>
-    <artifactId>assemblies</artifactId>
-    <version>8.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>pentaho-webpackage-archive-file-deployer</artifactId>
-
-  <packaging>feature</packaging>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.pentaho.webpackage</groupId>
-      <artifactId>archive-file-deployer-impl</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-  </dependencies>
-</project>
