+++ pentaho-webpackage/deployers/archive-deployer/src/main/java/org/pentaho/webpackage/deployer/archive/impl/UrlHandler.java
@@ -1,0 +1,31 @@
+/*
+ * This program is free software; you can redistribute it and/or modify it under the
+ * terms of the GNU Lesser General Public License, version 2.1 as published by the Free Software
+ * Foundation.
+ *
+ * You should have received a copy of the GNU Lesser General Public License along with this
+ * program; if not, you can obtain a copy at http://www.gnu.org/licenses/old-licenses/lgpl-2.1.html
+ * or from the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ *
+ * Copyright 2017 Pentaho Corporation. All rights reserved.
+ */
+
+package org.pentaho.webpackage.deployer.archive.impl;
+
+import org.osgi.service.url.AbstractURLStreamHandlerService;
+
+import java.io.IOException;
+import java.net.URL;
+import java.net.URLConnection;
+
+public class UrlHandler extends AbstractURLStreamHandlerService {
+  @Override public URLConnection openConnection( URL url ) throws IOException {
+    return new WebPackageURLConnection( new URL( url.getPath() ) );
+  }
+}
