+++ pentaho-webpackage/deployers/archive-deployer/src/test/resources/invalid.zip
