+++ pentaho-webpackage/deployers/archive-deployer/src/test/resources/my-simple-module-1.4.0.jar
