+++ pentaho-webpackage/deployers/archive-deployer/src/test/resources/no-package-json.jar
