+++ pentaho-webpackage/deployers/archive-deployer/src/test/resources/package.json
