+++ pentaho-webpackage/deployers/archive-deployer/src/test/resources/with-manifest.jar
