+++ pentaho-webpackage/extenders/http-extender/src/test/resources/invalid-package.json
