+++ pentaho-webpackage/extenders/http-extender/src/test/resources/pentaho-webpackage-2b-package.json
