+++ pentaho-webpackage/extenders/http-extender/src/test/resources/root-package.json
@@ -1,0 +1,4 @@
+{
+  "name": "package-on-root",
+  "version": "1.0"
+}
